% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popcontam.R
\name{new_pc_statevec}
\alias{new_pc_statevec}
\alias{validate_pc_statevec}
\alias{pc_statevec}
\title{Create a state vector for the population model}
\usage{
new_pc_statevec(state, statenames = names(state), t = 0)

validate_pc_statevec(statevec)

pc_statevec(state, statenames = names(state), t = 0)
}
\arguments{
\item{state}{Number of individuals in each state}

\item{statenames}{Name of each state (e.g. lifestages)}

\item{t}{Time}
}
\value{
A named vector with class \code{statevec}
}
\description{
Create a named vector representing the state of the population at a given
point in time
}
\section{Functions}{
\itemize{
\item \code{new_pc_statevec()}: Create the state vector

\item \code{validate_pc_statevec()}: Validate the state vector

}}
